+++ packages/cryptoplease/lib/bl/swap_tokens/transaction/swap_transaction_event.dart
@@ -6,6 +6,6 @@ class SwapTransactionEvent with _$SwapTransactionEvent {
     required Token inputToken,
     required Token outputToken,
     required Amount amount,
-    required double slippage,
+    required Decimal slippage,
   }) = SwapEvent;
 }
